refactor : 조회 dto 수정
